# utils: add hint to enable/disable object validity checks