@bjorn 2025-06-26
Btw, did you consider disabling these checks based on a hint (see [`SDL_SetHint`](https://wiki.libsdl.org/SDL3/SDL_SetHint))? That should make the speedup accessible to a lot more people, assuming the hint can be checked very fast (could probably be read only once). At least personally I tend to rely on system packages.