@mechakotik 2025-06-26
Looked into it again and it seems like SDL_CleanupTrays is not required and is a protection in case user forgets to call `SDL_DestroyTray` in the end.

> Btw, did you consider disabling these checks based on a hint (see [SDL_SetHint](https://wiki.libsdl.org/SDL3/SDL_SetHint))? That should make the speedup accessible to a lot more people, assuming the hint can be checked very fast (could probably be read only once). At least personally I tend to rely on system packages.

That's reasonable, thanks for the idea!