@bjorn 2025-06-27
> it seems like SDL_CleanupTrays is not required and is a protection in case user forgets to call `SDL_DestroyTray` in the end.

In that case this sounds like another case of hand-holding which makes life harder for people who're checking their code with a memory leak detector. It may then be acceptable that this cleanup no longer works when object validity checks are disabled, though it does seem like something that should be mentioned in the documentation of the hint.