@slouken 2025-08-25
I still want to compile out the validity checks entirely if someone builds with a flag to do so, and possibly disable parameter validation entirely, so I'm not sure whether this is the approach we want to take.