@mechakotik 2025-08-26
Do you want validity check to be enabled by default? If so, it is reasonable to support disabling them in runtime, because such environments as Linux distros and Steam Runtime use system-wide shared SDL library, and it's often painful to replace it with self-built version.

Checking hint value should have almost zero overhead, because checking value of non-changing variable can be easily handled by branch predictor. If this is not true, it's easy to support both compile- and run-time control of validity checks.